The initial commit :)
This initial commit works on my machine :)
This includes Sending a message from a stored procedure, run the database script, to the work, you need to run a worker, which work is pushed to from the manager that reports to the website.  The lighthouse discovers the services and allows you to force them to leave.
Rull all the project after running the sql to bulid the database and then put the test file in the project \ProcessorEastCoast\ExampleFile into the specified file. default should be c:\common\<locationname>.